+++ src/main/kotlin/com/cedricziel/idea/concourse/ConcourseIcons.kt
@@ -4,8 +4,8 @@ import com.intellij.icons.AllIcons
 import com.intellij.openapi.util.IconLoader
 
 object ConcourseIcons {
-    val CONCOURSE = IconLoader.getIcon("/META-INF/pluginIcon.svg")
-    val CONCOURSE_TASK = IconLoader.getIcon("/META-INF/pluginIcon.svg")
+    val CONCOURSE = IconLoader.getIcon("/META-INF/pluginIcon.svg", ConcourseIcons.javaClass)
+    val CONCOURSE_TASK = IconLoader.getIcon("/META-INF/pluginIcon.svg", ConcourseIcons.javaClass)
     val INPUT = AllIcons.Graph.ZoomIn
     val OUTPUT = AllIcons.Graph.ZoomOut
     val RESOURCE = AllIcons.Ide.Notification.Gear