+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 ### Added
 
 ### Changed
-
+- Migrate getResource usage to reference current classloader
+- Migrate getIcon usage to reference current classloader
+- Override getDisplayName for custom file types
 ### Deprecated
 
 ### Removed
@@ -21,7 +23,6 @@
 - Upgrade `io.gitlab.arturbosch.detekt` to 1.18.1
 - Upgrade `org.jlleitschuh.gradle.ktlint` to 10.2.0
 - Fix Ktlint issue
-
 ### Deprecated
 
 ### Removed